create nodes based on extension OR mediaType
